+++ packages/jiyeong-vanilla/README.md
@@ -27,11 +27,10 @@
 ### 구매 기능 구현
 
 - [x] 금액 표시창은 항상 숫자값을 표시하며 기본값은 0입니다.
-- [ ] 상품 버튼을 클릭했을 때 투입된 금액이 상품 가격보다 적으면 버튼을 누르는 동안 금액 표시창에 상품 금액이 표시됩니다.
-- [ ] 버튼을 누르지 않을 때는 투입된 금액이 표시됩니다.
+- [x] 상품 버튼을 클릭했을 때 투입된 금액이 상품 가격보다 적으면 버튼을 누르는 동안 금액 표시창에 상품 금액이 표시됩니다.
+- [x] 버튼을 누르지 않을 때는 투입된 금액이 표시됩니다.
 - [ ] 금액 표시(투입금액 총합 - 상품금액)으로 나타내야함
-- [ ] 구매 취소는 없음
-- [ ] 상품 버튼을 클릭했을 때 투입된 금액이 상품 가격보다 적으면 버튼을 누르는 동안 금액 표시창에 상품 금액이 표시됩니다.
+- [x] 구매 취소는 없음
 
 ### 반환 기능 구현
